requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0
jdatetime>=4.1.0

# Reporting
reportlab>=4.0.0
markdown>=3.5.0
pypdf>=3.17.0

# Alerts
python-telegram-bot>=20.0
aiohttp>=3.9.0

# Database migrations
alembic>=1.13.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
click>=8.1.0
pyyaml>=6.0.1
